backend/requirements.txt
websockets>=12.0
sounddevice>=0.4.6
numpy>=1.24.0
python-dotenv>=1.0.0
snowflake-connector-python>=3.9.0
